Bot-Tele/requirements.txt
telethon>=1.36.0
python-telegram-bot>=20.7
google-genai>=1.0.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
aiosqlite>=0.20.0
